DEV-005: fix Converters.kt compile error (Long vs Int epoch-day type)
CI run 34986735142 (commit 4e049b4) failed to compile :core:database:
"Return type mismatch: expected 'Int?', actual 'Long?'" at
Converters.kt:30. kotlinx-datetime 0.8.0's LocalDate.toEpochDays()
returns Long, not Int as the converter assumed — confirmed by extracting
and javap-inspecting the actual cached kotlinx-datetime-jvm-0.8.0.jar in
this sandbox (already resolved locally via :core:common's own dependency
on it), rather than assumed from memory.

Fixed localDateToEpochDay/epochDayToLocalDate to use Long throughout,
matching the Instant epoch-millis columns' own type.

+++ core/database/src/main/kotlin/com/stepwise/core/database/Converters.kt
@@ -27,10 +27,10 @@ class Converters {
     fun epochMillisToInstant(value: Long?): Instant? = value?.let { Instant.fromEpochMilliseconds(it) }
 
     @TypeConverter
-    fun localDateToEpochDay(value: LocalDate?): Int? = value?.toEpochDays()
+    fun localDateToEpochDay(value: LocalDate?): Long? = value?.toEpochDays()
 
     @TypeConverter
-    fun epochDayToLocalDate(value: Int?): LocalDate? = value?.let { LocalDate.fromEpochDays(it) }
+    fun epochDayToLocalDate(value: Long?): LocalDate? = value?.let { LocalDate.fromEpochDays(it) }
 
     @TypeConverter
     fun localTimeToSecondOfDay(value: LocalTime?): Int? =
